requirements.txt
requests>=2.28.0
requests-pkcs12>=1.0
python-dotenv>=1.0.0
websockets>=12.0
pycryptodome>=3.19
